requirements.txt
# Core
python-dotenv
MetaTrader5
numpy
pandas
Pillow
certifi
scikit-learn

# UI
matplotlib
